add monitoring test
