+++ testing/integration/beat_receivers_test.go
@@ -7,24 +7,35 @@
 package integration
 
 import (
+	"bytes"
 	"context"
 	"encoding/base64"
 	"fmt"
 	"io"
 	"net/http"
+	"strings"
 	"testing"
+	"text/template"
 	"time"
 
 	"gopkg.in/yaml.v2"
 
 	"github.com/elastic/elastic-agent-libs/kibana"
 	"github.com/elastic/elastic-agent-libs/testing/estools"
+	"github.com/elastic/elastic-agent/internal/pkg/agent/application/paths"
 	atesting "github.com/elastic/elastic-agent/pkg/testing"
 	"github.com/elastic/elastic-agent/pkg/testing/define"
 	"github.com/elastic/elastic-agent/pkg/testing/tools/testcontext"
+	"github.com/elastic/elastic-agent/pkg/utils"
 
 	"github.com/gofrs/uuid/v5"
 	"github.com/stretchr/testify/require"
+)
+
+var (
+	logsEADocs       estools.Documents
+	logsOTelDocs     estools.Documents
+	commonLogMessage = "Determined allowed capabilities"
 )
 
 // TestAgentMonitoring is a test to provide a baseline for what
@@ -151,6 +162,7 @@ func TestAgentMonitoring(t *testing.T) {
 		{dsType: "logs", dsDataset: "elastic_agent", dsNamespace: info.Namespace},
 		{dsType: "metrics", dsDataset: "elastic_agent.elastic_agent", dsNamespace: info.Namespace},
 		{dsType: "metrics", dsDataset: "elastic_agent.filebeat", dsNamespace: info.Namespace},
+		{dsType: "metrics", dsDataset: "elastic_agent.filebeat_input", dsNamespace: info.Namespace},
 		{dsType: "metrics", dsDataset: "elastic_agent.metricbeat", dsNamespace: info.Namespace},
 	}
 
@@ -183,4 +195,288 @@ func TestAgentMonitoring(t *testing.T) {
 			2*time.Minute, 5*time.Second,
 			"No documents found for type: %s, dataset: %s, namespace: %s", tc.dsType, tc.dsDataset, tc.dsNamespace)
 	}
+
+	// get self monitoring logs
+	var agentDocs estools.Documents
+	fbMonitoringIndex := "logs-elastic_agent-" + info.Namespace
+	require.Eventually(t,
+		func() bool {
+			findCtx, findCancel := context.WithTimeout(ctx, 10*time.Second)
+			defer findCancel()
+
+			agentDocs, err = estools.GetLogsForIndexWithContext(findCtx, info.ESClient, ".ds-"+fbMonitoringIndex+"*", map[string]interface{}{
+				"message": commonLogMessage,
+			})
+			require.NoError(t, err)
+
+			if agentDocs.Hits.Total.Value != 0 {
+				logsEADocs = agentDocs
+				return true
+			}
+			return false
+		},
+		2*time.Minute, 1*time.Second, "could not find monitoring log")
+
+}
+
+func TestMonitoringOTelLogs(t *testing.T) {
+	require.NotPanics(
+		t, func() {
+			_ = logsEADocs.Hits.Hits[0].Source
+		}, "monitoring logs from elastic-agent should exist before proceeding",
+	)
+	// This test compares the docs ingested with self monitoring in otel mode
+	info := define.Require(t, define.Requirements{
+		Group: Default,
+		Local: true,
+		OS: []define.OS{
+			{Type: define.Linux},
+			{Type: define.Darwin},
+		},
+		Stack: &define.Stack{},
+	})
+
+	fbReceiverMonitoringIndex := "logs-elastic_agent-monitoringotel"
+
+	type configOptions struct {
+		InputPath      string
+		HomeDir        string
+		ESEndpoint     string
+		ESApiKey       string
+		BeatsESApiKey  string
+		SocketEndpoint string
+	}
+	esEndpoint, err := getESHost()
+	require.NoError(t, err, "error getting elasticsearch endpoint")
+	esApiKey, err := createESApiKey(info.ESClient)
+	require.NoError(t, err, "error creating API key")
+	require.NotEmptyf(t, esApiKey.Encoded, "api key is invalid %q", esApiKey)
+
+	// Start monitoring in otel mode
+	fixture, err := define.NewFixtureFromLocalBuild(t, define.Version())
+	require.NoError(t, err)
+
+	ctx, cancel := testcontext.WithDeadline(t, context.Background(), time.Now().Add(5*time.Minute))
+	defer cancel()
+
+	err = fixture.Prepare(ctx)
+	require.NoError(t, err)
+
+	// Get the running dir
+	inputPath, err := fixture.GetRunningDir(ctx)
+	socketEndpoint := utils.SocketURLWithFallback(uuid.Must(uuid.NewV4()).String(), paths.TempDir())
+
+	configTemplateOTel := `
+agent.grpc:
+  port: 6792	
+receivers:
+  filebeatreceiver/filestream-monitoring:
+    filebeat:
+      inputs:
+        - type: filestream
+          enabled: true
+          id: filestream-monitoring-agent
+          paths:
+            - {{.InputPath}}/data/elastic-agent-*/logs/elastic-agent-*.ndjson
+            - {{.InputPath}}/data/elastic-agent-*/logs/elastic-agent-watcher-*.ndjson
+          close:
+            on_state_change:
+              inactive: 5m	  
+          parsers:
+            - ndjson:
+                add_error_key: true
+                message_key: message
+                overwrite_keys: true
+                target: ""
+          processors:
+            - add_fields:
+                fields:
+                  dataset: elastic_agent
+                  namespace: monitoringotel
+                  type: logs
+                target: data_stream
+            - add_fields:
+                fields:
+                  dataset: elastic_agent
+                target: event
+            - add_fields:
+                fields:
+                  id: 0ddca301-e7c0-4eac-8432-7dd05bc9cb06
+                  snapshot: false
+                  version: 8.19.0
+                target: elastic_agent
+            - add_fields:
+                fields:
+                  id: 0879f47d-df41-464d-8462-bc2b8fef45bf
+                target: agent
+            - drop_event:
+                when:
+                  regexp:
+                    name: .*-monitoring$
+            - drop_event:
+                when:
+                  regexp:
+                    message: ^Non-zero metrics in the last
+            - copy_fields:
+                fields:
+                  - from: data_stream.dataset
+                    to: data_stream.dataset_original
+            - drop_fields:
+                fields:
+                  - data_stream.dataset
+            - copy_fields:
+                fail_on_error: false
+                fields:
+                  - from: component.dataset
+                    to: data_stream.dataset
+                ignore_missing: true
+            - copy_fields:
+                fail_on_error: false
+                fields:
+                  - from: data_stream.dataset_original
+                    to: data_stream.dataset
+            - drop_fields:
+                fields:
+                  - data_stream.dataset_original
+                  - event.dataset
+            - copy_fields:
+                fields:
+                  - from: data_stream.dataset
+                    to: event.dataset
+            - drop_fields:
+                fields:
+                  - ecs.version
+                ignore_missing: true
+    output:
+      otelconsumer:
+    queue:
+      mem:
+        flush:
+          timeout: 0s
+    logging:
+      level: info
+      selectors:
+        - '*'
+    filebeat.config.modules.enabled: false
+    http.enabled: true
+    http.host: {{ .SocketEndpoint }}
+exporters:
+  debug:
+    use_internal_logger: false
+    verbosity: detailed
+  elasticsearch/log:
+    endpoints:
+      - {{.ESEndpoint}}
+    compression: none
+    api_key: {{.ESApiKey}}
+    logs_dynamic_index:
+      enabled: true
+    batcher:
+      enabled: true
+      flush_timeout: 0s
+    mapping:
+      mode: bodymap	  
+service:
+  pipelines:
+    logs:
+      receivers:
+        - filebeatreceiver/filestream-monitoring
+      exporters:
+        - elasticsearch/log  
+`
+
+	var configBuffer bytes.Buffer
+	template.Must(template.New("config").Parse(configTemplateOTel)).Execute(&configBuffer,
+		configOptions{
+			InputPath:      inputPath,
+			ESEndpoint:     esEndpoint,
+			ESApiKey:       esApiKey.Encoded,
+			SocketEndpoint: socketEndpoint,
+		})
+	configOTelContents := configBuffer.Bytes()
+	t.Cleanup(func() {
+		if t.Failed() {
+			t.Logf("Contents of agent config file:\n%s\n", string(configOTelContents))
+		}
+	})
+
+	// configures, starts and waits for elastic-agent to be healthy
+	err = fixture.Configure(ctx, configOTelContents)
+	require.NoError(t, err)
+
+	cmd, err := fixture.PrepareAgentCommand(ctx, nil)
+	require.NoError(t, err)
+	cmd.WaitDelay = 1 * time.Second
+
+	var output strings.Builder
+	cmd.Stderr = &output
+	cmd.Stdout = &output
+
+	err = cmd.Start()
+	require.NoError(t, err)
+
+	t.Cleanup(func() {
+		if t.Failed() {
+			t.Log("Elastic-Agent output:")
+			t.Log(output.String())
+		}
+	})
+
+	require.Eventually(t, func() bool {
+		err = fixture.IsHealthy(ctx)
+		if err != nil {
+			t.Logf("waiting for agent healthy: %s", err.Error())
+			return false
+		}
+		return true
+	}, 30*time.Second, 1*time.Second)
+
+	var otelDocs estools.Documents
+	require.Eventually(t,
+		func() bool {
+			findCtx, findCancel := context.WithTimeout(ctx, 10*time.Second)
+			defer findCancel()
+
+			otelDocs, err = estools.GetLogsForIndexWithContext(findCtx, info.ESClient, ".ds-"+fbReceiverMonitoringIndex+"*", map[string]interface{}{
+				"message": commonLogMessage,
+			})
+			require.NoError(t, err)
+
+			if otelDocs.Hits.Total.Value != 0 {
+				logsOTelDocs = otelDocs
+				return true
+			}
+			return false
+		},
+		2*time.Minute, 1*time.Second, "could not find otel monitoring log")
+
+	// Stop monitoring agent otel mode
+	cancel()
+	cmd.Wait()
+
+	agent := logsEADocs.Hits.Hits[0].Source
+	otel := logsOTelDocs.Hits.Hits[0].Source
+	ignoredFields := []string{
+		// Expected to change between agentDocs and OtelDocs
+		"@timestamp",
+		"agent.ephemeral_id",
+		"agent.id",
+		"agent.version",
+		"data_stream.namespace",
+		"log.file.inode",
+		"log.file.fingerprint",
+		"log.file.path",
+		"log.offset",
+
+		// needs investigation
+		"event.agent_id_status",
+		"event.ingested",
+
+		// elastic_agent * fields are hardcoded in processor list for now which is why they differ
+		"elastic_agent.id",
+		"elastic_agent.snapshot",
+		"elastic_agent.version",
+	}
+
+	AssertMapsEqual(t, agent, otel, ignoredFields, "expected documents to be equal")
 }
